pyproject.toml
[project]
name = "python-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.4",
  "asyncpg>=0.30.0",
  "asyncpg-stubs>=0.30.2",
  "bcrypt>=4.3.0",
  "discord-py>=2.5.2",
  "jishaku>=2.6.0",
  "psycopg2-binary>=2.9.10",
  "pydantic-settings>=2.10.1",
  "python-dotenv>=1.1.1",
  "pytz>=2025.2",
  "redis[hiredis]>=6.2.0",
  "sqlalchemy>=2.0.41",
]

[tool.pyright]
venv = ".venv"
venvPath = "."
typeCheckingMode = "strict"
reportArgumentType = "none"
# reportMissingTypeStubs = "none"
# reportUnknownVariableType = "none"
# reportUnknownMemberType = "none"
# reportAttributeAccessIssue = "none"
include = ["."]

[dependency-groups]
dev = ["alembic>=1.16.4", "pyright>=1.1.403", "ruff>=0.12.5"]
